# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponse_RunComparisonView_": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceView"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "$ref": "#/components/schemas/RunComparisonView"
          },
          "envelope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ResponseScope"
          },
          "stage_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Detail"
          },
          "stage_status": {
            "$ref": "#/components/schemas/StageStatusView"
          },
          "standing": {
            "$ref": "#/components/schemas/GraphStandingView"
          },
          "standing_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standing Notice"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingMeta"
          }
        },
        "required": [
          "scope",
          "provenance",
          "standing",
          "stage_status",
          "timing",
          "generated_at",
          "data"
        ],
        "title": "ApiResponse[RunComparisonView]",
        "type": "object"
      },
      "ApiResponse_Union_RunView__NoneType__": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceView"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunView"
              },
              {
                "type": "null"
              }
            ]
          },
          "envelope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ResponseScope"
          },
          "stage_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Detail"
          },
          "stage_status": {
            "$ref": "#/components/schemas/StageStatusView"
          },
          "standing": {
            "$ref": "#/components/schemas/GraphStandingView"
          },
          "standing_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standing Notice"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingMeta"
          }
        },
        "required": [
          "scope",
          "provenance",
          "standing",
          "stage_status",
          "timing",
          "generated_at",
          "data"
        ],
        "title": "ApiResponse[Union[RunView, NoneType]]",
        "type": "object"
      },
...
      "ApiResponse_tuple_RunView__________": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceView"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/RunView"
            },
            "title": "Data",
            "type": "array"
          },
          "envelope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "title": "Evidence",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ResponseScope"
          },
          "stage_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Detail"
          },
          "stage_status": {
            "$ref": "#/components/schemas/StageStatusView"
          },
          "standing": {
            "$ref": "#/components/schemas/GraphStandingView"
          },
          "standing_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standing Notice"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingMeta"
          }
        },
        "required": [
          "scope",
          "provenance",
          "standing",
          "stage_status",
          "timing",
          "generated_at",
          "data"
        ],
        "title": "ApiResponse[tuple[RunView, ...]]",
        "type": "object"
      },
...
      "ConfidenceView": {
        "additionalProperties": false,
        "description": "A confidence number that cannot exist without its decomposition.\n\n`components` carries `min_length=1`, so **the API is structurally incapable of\nreturning a bare float wearing a confidence's name**. LAW-EVIDENCE says a bare float is\na defect; this makes it an unconstructable one rather than something a test has to go\nlooking for. It is the same move ADR-0075 made for a recommendation's evidence, in the\nplace a client actually reads.\n\n`scalar` is derived and names the `aggregation` that produced it, exactly as\n`ConfidenceVector` does. A consumer that reads `scalar` and ignores `components` has\nbeen given everything it needs to know better.",
        "properties": {
          "aggregation": {
            "minLength": 1,
            "title": "Aggregation",
            "type": "string"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/ConfidenceComponentView"
            },
            "minItems": 1,
            "title": "Components",
            "type": "array"
          },
          "provenance_class": {
            "$ref": "#/components/schemas/ProvenanceClass"
          },
          "scalar": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Scalar",
            "type": "number"
          }
        },
        "required": [
          "scalar",
          "aggregation",
          "provenance_class",
          "components"
        ],
        "title": "ConfidenceView",
        "type": "object"
      },
...
      "EvidenceReference": {
        "additionalProperties": false,
        "description": "A citation resolving back to the record that justified an assertion.\n\nA reference, not the evidence itself: `CONVENTIONS.md` \u00a77 forbids any message or\npayload from carrying a raw source record, and `source_locator` names where the record\nis without reproducing what it says.",
        "properties": {
          "dataset_version": {
            "minLength": 1,
            "title": "Dataset Version",
            "type": "string"
          },
          "evidence_record_id": {
            "minLength": 1,
            "title": "Evidence Record Id",
            "type": "string"
          },
          "source_locator": {
            "minLength": 1,
            "title": "Source Locator",
            "type": "string"
          }
        },
        "required": [
          "evidence_record_id",
          "dataset_version",
          "source_locator"
        ],
        "title": "EvidenceReference",
        "type": "object"
      },
...
      "GraphStandingView": {
        "description": "Which graph a traversal walked (ADR-0059), mirrored for the presentation layer.\n\nA mirror rather than a re-export because `GraphStanding` lives in `causal_engine` (L6)\nand the API may not import it. The two are kept in step by\n`tests/api/test_standing_mirror_is_total.py`, which enumerates both enums: a member\nadded to one and not the other fails, so the mirror cannot silently narrow.\n\n`UNPROMOTED_DIAGNOSTIC` is carried into responses ON PURPOSE. It marks output the\nengine has disowned -- walked over module 10's scored graph before promotion -- and\nADR-0059 defends it with a required field, one construction site, a validator refusing\n`INFERRED`, and a fixed notice. OQ-026 records the residual risk honestly: none of\nthose four mechanisms survives a screenshot. The API adds the fifth it can: the notice\nis a REQUIRED field on any response carrying this standing, not an optional annotation.",
        "enum": [
          "STATED",
          "UNPROMOTED_DIAGNOSTIC"
        ],
        "title": "GraphStandingView",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OutputEnvelope": {
        "additionalProperties": false,
        "description": "Carried by every persisted artifact and every API response body.\n\n`CONVENTIONS.md` \u00a711: an output without its envelope cannot be verified and is a\ndefect. `execution_id` is present for correlation and is excluded from determinism\ndiffs by `scripts/check_determinism.py`.",
        "properties": {
          "dataset_version": {
            "title": "Dataset Version",
            "type": "string"
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "execution_id": {
            "title": "Execution Id",
            "type": "string"
          },
          "graph_projection_version": {
            "title": "Graph Projection Version",
            "type": "string"
          },
          "ontology_hash": {
            "title": "Ontology Hash",
            "type": "string"
          },
          "ontology_version": {
            "title": "Ontology Version",
            "type": "string"
          },
          "rule_pack_version": {
            "title": "Rule Pack Version",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "ontology_version",
          "ontology_hash",
          "dataset_version",
          "rule_pack_version",
          "engine_version",
          "graph_projection_version",
          "seed",
          "execution_id"
        ],
        "title": "OutputEnvelope",
        "type": "object"
      },
...
      "ProvenanceClass": {
        "description": "How a thing came to be known. Orthogonal to how confident we are in it.",
        "enum": [
          "OBSERVED",
          "ASSUMED",
          "STATISTICAL",
          "INFERRED",
          "SIMULATED"
        ],
        "title": "ProvenanceClass",
        "type": "string"
      },
      "ProvenanceSummary": {
        "additionalProperties": false,
        "description": "The epistemic status of a payload, without flattening the five classes.\n\n`weakest` is `core.provenance.combine` over everything in the payload, which returns\nthe weakest input and never a stronger one (ADR-0005). `present` is every class the\npayload actually contains, so a reader can see that a response mixing `OBSERVED` facts\nwith one `SIMULATED` value is not uniformly simulated.\n\nBoth are carried because either alone misleads. `weakest` alone hides that most of a\npayload is observed; `present` alone leaves a reader to compute the floor themselves\nand get it wrong. `docs/architecture.md` \u00a72 forbids the API from \"flattening five\nprovenance classes into two\", and reporting only one summary number would be that.",
        "properties": {
          "present": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceClass"
            },
            "minItems": 1,
            "title": "Present",
            "type": "array"
          },
          "weakest": {
            "$ref": "#/components/schemas/ProvenanceClass"
          }
        },
        "required": [
          "weakest",
          "present"
        ],
        "title": "ProvenanceSummary",
        "type": "object"
      },
...
      "ResponseScope": {
        "description": "Whether a body describes engine output or the deployment around it.\n\nThis distinction exists because of an honest problem with requiring an envelope\neverywhere: an `OutputEnvelope` describes ONE run, and a handful of endpoints are not\nabout one run. A listing of runs spans many. A job that refused before its packs\nresolved has none yet -- migration 0017's `pipeline_job.run_id` is nullable for exactly\nthat reason, and minting a placeholder would be an identifier that addresses nothing\n(`CONVENTIONS.md` \u00a79).\n\nThe alternative was to synthesize an envelope of empty strings for those, which would\nmean a body carrying version fields that describe nothing. `CONVENTIONS.md` \u00a711 wants\noutputs to be VERIFIABLE; a fabricated envelope is worse than an absent one, because it\nis unverifiable while looking verified.\n\nSo the scope is declared, and the validator below makes `RUN_SCOPED` mean what it says.\n`CATALOG` is a visible choice in a diff, the set of routes using it is small, and\n`tests/api/test_catalog_scope_is_closed.py` fails if a route adopts it without being on\nthe declared list.",
        "enum": [
          "RUN_SCOPED",
          "CATALOG"
        ],
        "title": "ResponseScope",
        "type": "string"
      },
...
      "RunComparisonView": {
        "additionalProperties": false,
        "description": "Two runs and the inputs that differ, so one input's effect can be isolated.\n\n`differing_inputs` is the point of the endpoint. Two runs that differ in one `RunKey`\nfield isolate that field's effect on the output (ADR-0013's stated positive\nconsequence); two runs that differ in four do not, and a comparison that did not say\nwhich had changed would invite the first reading regardless.",
        "properties": {
          "differing_inputs": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Differing Inputs",
            "type": "array"
          },
          "identical": {
            "title": "Identical",
            "type": "boolean"
          },
          "left": {
            "$ref": "#/components/schemas/RunView"
          },
          "right": {
            "$ref": "#/components/schemas/RunView"
          }
        },
        "required": [
          "left",
          "right",
          "identical"
        ],
        "title": "RunComparisonView",
        "type": "object"
      },
      "RunView": {
        "additionalProperties": false,
        "description": "One Run: the five inputs that determine it, and what has been derived from it.",
        "properties": {
          "causal_edge_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Causal Edge Count",
            "type": "integer"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "dataset_version": {
            "minLength": 1,
            "title": "Dataset Version",
            "type": "string"
          },
          "engine_version": {
            "minLength": 1,
            "title": "Engine Version",
            "type": "string"
          },
          "graph_projection_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Projection Version"
          },
          "ontology_hash": {
            "minLength": 1,
            "title": "Ontology Hash",
            "type": "string"
          },
          "rule_pack_version": {
            "minLength": 1,
            "title": "Rule Pack Version",
            "type": "string"
          },
          "run_id": {
            "minLength": 1,
            "title": "Run Id",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "dataset_version",
          "ontology_hash",
          "rule_pack_version",
          "engine_version",
          "seed"
        ],
        "title": "RunView",
        "type": "object"
      },
...
      "StageStatusView": {
        "description": "Whether the pipeline stage behind a payload actually produced it.\n\n`NOT_RUNNABLE` is what an endpoint returns when its owning module has not been built --\nmodules 7, 8 and 15 today. It is a 200 with a full envelope and an explicit statement,\nnever an empty success: `docs/architecture.md` \u00a72 requires that a partially computed\nrun \"returns what exists with the stage stated, never a silently truncated graph\", and\nan empty list with no explanation is exactly the silent truncation that forbids.",
        "enum": [
          "COMPLETE",
          "PARTIAL",
          "NOT_RUNNABLE"
        ],
        "title": "StageStatusView",
        "type": "string"
      },
...
      "TimingMeta": {
        "additionalProperties": false,
        "description": "What one operation cost, and what it was allowed to cost.\n\nCarried on every API response. `CONVENTIONS.md` \u00a711 excludes performance measurements\nfrom determinism comparisons, so this whole model is excluded -- which is why the\nbudget and its name travel WITH the measurement rather than being looked up by a\nreader: a number the determinism diff ignores must still be interpretable on its own.\n\n`within_budget` is None when the operation has no declared budget. That is a third\nstate and not a quiet `True`: \"this was fast enough\" and \"nobody said how fast this\nshould be\" are different claims, and collapsing them would let an unbudgeted endpoint\nreport compliance it was never measured for.",
        "properties": {
          "budget_enforced": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Enforced"
          },
          "budget_name": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetName"
              },
              {
                "type": "null"
              }
            ]
          },
          "budget_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Seconds"
          },
          "elapsed_seconds": {
            "title": "Elapsed Seconds",
            "type": "number"
          },
          "operation": {
            "title": "Operation",
            "type": "string"
          },
          "within_budget": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Within Budget"
          }
        },
        "required": [
          "operation",
          "elapsed_seconds"
        ],
        "title": "TimingMeta",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "A domain-agnostic causal intelligence engine. Every response carries its run, its provenance, its evidence references and its confidence decomposition; no endpoint returns a confidence without its components.",
    "title": "CausaLog",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/runs": {
      "get": {
        "description": "List every run materialized in this process, in canonical sequence.\n\nCATALOG scope: this body spans many runs, so there is no single `OutputEnvelope` that\ndescribes it. Each ITEM carries the five inputs that identify its run; synthesizing an\nenvelope for the listing would be fabricating version fields that describe nothing.",
        "operationId": "list_runs_v1_runs_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_tuple_RunView__________"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/compare": {
      "get": {
        "description": "Compare two runs and name the inputs that differ.",
        "operationId": "compare_runs_v1_runs_compare_get",
        "parameters": [
          {
            "in": "query",
            "name": "left",
            "required": true,
            "schema": {
              "title": "Left",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "right",
            "required": true,
            "schema": {
              "title": "Right",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RunComparisonView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare Runs",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/runs/{run_id}/inferred-artifacts": {
      "delete": {
        "description": "Drop a run's inferred artifacts. Observed facts are untouched and unreachable here.",
        "operationId": "delete_inferred_artifacts_v1_runs__run_id__inferred_artifacts_delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Union_RunView__NoneType__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Inferred Artifacts",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}